use String constructor directly for short ascii values
On jdk9+ the String internals changed and it is more efficient
to just use the String contructor that takes a byte array when
dealing with ASCII values. Since jdk11 is the most recent LTS
version with these changes and there are some methods added to
String that can be used to detect it, this change uses the
String constructor directly when on a newer JDK. The old
behavior is preserved for jdk8.