Add support for CBOR stringref extension. (#347)
Add support for parsing CBOR with stringref extension.

See http://cbor.schmorp.de/stringref for stringref definition.
When a stringref namespace is encountered, indices will be generated for
each valid binary and text string and used when referenced.

Nested tags are now supported. The previous getCurrentTag() function is
still supported for backwards compatibility, and returns the first tag.
Otherwise a new TagList type is used to efficiently handle multiple tags.

* Add support for writing CBOR with stringref extension.

When enabled with the STRINGREF feature, a namespace is created for the
root element and indices are produced for valid binary and text strings.

* Remove tag copying for CBOR event copies.

Copying tags can be dangerous in two main cases:
1. Jackson creates its own tags, such as for BigNum or the self describe
   tag.
2. The semantic meaning of an element changes. For example, converting
   between stringrefs and no stringrefs.

In some cases it would result in duplicate tags, while in others it may
cause completely incorrect tags, such as a stringref tag applying to a
string rather than an integer. Before nested tag support was added this
could generate files that Jackson could not parse, though even with nested
tag support it may cause undefined behavior.