Add WRITE_MINIMAL_DOUBLES feature to CBORGenerator (#356)
Enabling WRITE_MINIMAL_DOUBLES works similarly to WRITE_MINIMAL_INTS,
choosing to use a smaller data representation (single-precision float) when
the conversion will result in no loss of data. This can provide a
substantial savings for data that contains many doubles that don't require
full precision.

Cleaned up some of the writing code for floats and doubles to avoid code
duplication, and reserve the proper amount of space for output for both
types.